Remove utils_ramdya_lab.py
