+++ deepfly/core.py
@@ -1,8 +1,9 @@
 import os.path
 import math  # inf
 import numpy as np
+import re
+import deepfly.logger as logger
 
-from deepfly.utils_ramdya_lab import find_default_camera_ordering
 from deepfly.CameraNetwork import CameraNetwork
 from deepfly.Config import config
 from deepfly.os_util import write_camera_order, read_camera_order, read_calib, get_max_img_id
@@ -17,6 +18,19 @@ from deepfly import logger
 
 from sklearn.neighbors import NearestNeighbors
 import pickle
+
+known_users = [  
+    (r'/CLC/', [0, 6, 5, 4, 3, 2, 1]),
+    (r'data/test', [0, 1, 2, 3, 4, 5, 6])
+]
+
+
+def find_default_camera_ordering(input_folder):
+    input_folder = str(input_folder)  # allows input_folder to be a pathlib.Path instance
+    candidates = [ordering for (regex, ordering) in known_users if re.search(regex, input_folder)]
+    if candidates:
+        logger.debug(f'Default camera ordering found for current user: {candidates[0]}')
+        return np.array(candidates[0])
 
 
 class Core:
